+++ DarkCodex/Kineticist.cs
@@ -18,6 +18,20 @@ using static Kingmaker.Visual.Animation.Kingmaker.Actions.UnitAnimationActionCas
 using Kingmaker.ElementsSystem;
 using Kingmaker.UnitLogic.Abilities.Components;
 using Kingmaker.Designers.EventConditionActionSystem.Actions;
+using Kingmaker.UnitLogic.Commands;
+using Kingmaker.EntitySystem.Stats;
+using Kingmaker.Enums.Damage;
+using Kingmaker.RuleSystem.Rules.Damage;
+using Kingmaker.UnitLogic.Mechanics.Actions;
+using Kingmaker.UnitLogic.Mechanics;
+using Kingmaker.RuleSystem;
+using Kingmaker.Enums;
+using Kingmaker.UnitLogic.Abilities;
+using Kingmaker.UnitLogic.Mechanics.Components;
+using Kingmaker.Blueprints.Classes.Spells;
+using Kingmaker.UnitLogic.Abilities.Components.Base;
+using Kingmaker.ResourceLinks;
+using Kingmaker.Blueprints.Facts;
 
 namespace DarkCodex
 {
@@ -25,7 +39,7 @@ namespace DarkCodex
     {
         public static void createKineticistBackground()
         {
-            var x = Helper.CreateBlueprintFeature(
+            var feature = Helper.CreateBlueprintFeature(
                 "BackgroundElementalist",
                 "Elemental Plane Outsider",
                 "Elemental Plane Outsider count as 1 level higher for determining Kineticist levels.",
@@ -34,7 +48,31 @@ namespace DarkCodex
                 0,
                 Helper.CreateClassLevelsForPrerequisites("42a455d9ec1ad924d889272429eb8391", 1)); //kineticist class
 
-            Helper.AppendAndReplace(ref ResourcesLibrary.TryGetBlueprint<BlueprintFeatureSelection>("fa621a249cc836f4382ca413b976e65e").m_AllFeatures, x.ToRef());
+            Helper.AppendAndReplace(ref ResourcesLibrary.TryGetBlueprint<BlueprintFeatureSelection>("fa621a249cc836f4382ca413b976e65e").m_AllFeatures, feature.ToRef());
+        }
+
+        public static void createExtraWildTalentFeat(bool enabled = true)
+        {
+            var kineticist_class = ResourcesLibrary.TryGetBlueprint<BlueprintCharacterClass>("42a455d9ec1ad924d889272429eb8391");
+            var infusion_selection = ResourcesLibrary.TryGetBlueprint<BlueprintFeatureSelection>("58d6f8e9eea63f6418b107ce64f315ea");
+            var wildtalent_selection = ResourcesLibrary.TryGetBlueprint<BlueprintFeatureSelection>("5c883ae0cd6d7d5448b7a420f51f8459");
+
+            var extra_wild_talent_selection = Helper.CreateBlueprintFeatureSelection(
+                "ExtraWildTalentFeat",
+                "Extra Wild Talent",
+                "You gain a wild talent for which you meet the prerequisites. You can select an infusion or a non-infusion wild talent, but not a blast or defense wild talent.\nSpecial: You can take this feat multiple times. Each time, you must choose a different wild talent.",
+                null,
+                ResourcesLibrary.TryGetBlueprint<BlueprintFeature>("42f96fc8d6c80784194262e51b0a1d25").Icon, //ExtraArcanePool.Icon
+                FeatureGroup.Feat,
+                Helper.CreatePrerequisiteClassLevel(kineticist_class, 1, true)
+            );
+            extra_wild_talent_selection.Ranks = 10;
+
+            extra_wild_talent_selection.m_AllFeatures = Helper.Append(infusion_selection.m_AllFeatures,     //InfusionSelection
+                                                                    wildtalent_selection.m_AllFeatures);  //+WildTalentSelection
+
+            if (enabled)
+                Helper.AddFeats(extra_wild_talent_selection);
         }
 
         // known issue:
@@ -51,11 +89,11 @@ namespace DarkCodex
 
             // rename buffs, so it's easier to tell them apart
             buff1.m_Icon = gather_original_ab.Icon;
-            buff1.m_DisplayName = Helper.CreateString(buff1.m_DisplayName + " Lv1");
+            buff1.m_DisplayName = Helper.CreateString("Gather Power Lv1");
             buff2.m_Icon = gather_original_ab.Icon;
-            buff2.m_DisplayName = Helper.CreateString(buff2.m_DisplayName + " Lv2");
+            buff2.m_DisplayName = Helper.CreateString("Gather Power Lv2");
             buff3.m_Icon = gather_original_ab.Icon;
-            buff3.m_DisplayName = Helper.CreateString(buff3.m_DisplayName + " Lv3");
+            buff3.m_DisplayName = Helper.CreateString("Gather Power Lv3");
 
             // new buff that halves movement speed, disallows normal gathering
             var mobile_debuff = Helper.CreateBlueprintBuff(
@@ -74,7 +112,6 @@ namespace DarkCodex
             gather_original_ab.AddComponents(Helper.CreateAbilityRequirementHasBuffs(true, mobile_debuff));
 
             // ability as free action that applies buff and 1 level of gatherpower
-            // - is free action
             // - increases gather power by 1 level, similiar to GatherPower:6dcbffb8012ba2a4cb4ac374a33e2d9a
             // - applies debuff
             // - get same restriction as usual gathering
@@ -103,12 +140,12 @@ namespace DarkCodex
             // same as above but standard action and 2 levels of gatherpower
             var one2three = Helper.CreateConditional(Helper.CreateContextConditionHasBuff(buff1).ObjToArray(), new GameAction[] { Helper.CreateContextActionRemoveBuff(buff1), Helper.CreateContextActionApplyBuff(buff3, 2) });
             var zero2two = Helper.CreateConditional(Helper.CreateConditionHasNoBuff(buff1, buff2, buff3), new GameAction[] { Helper.CreateContextActionApplyBuff(buff2, 2) });
-            var hasMoveAction = Helper.CreateAbilityRequirementActionAvailable(false, ActionType.Move, 4.5f);
+            var hasMoveAction = Helper.CreateAbilityRequirementActionAvailable(false, ActionType.Move, 6f);
             var lose_halfmove = new ContextActionUndoAction(command: UnitCommand.CommandType.Move, amount: -1.5f);
             var mobile_gathering_long_ab = Helper.CreateBlueprintAbility(
                 "MobileGatheringLong",
                 "Mobile Gathering (Full Round)",
-                "You may move up to half your normal speed while gathering power.\nNote: Use half your speed before using this.",
+                "You may move up to half your normal speed while gathering power.",
                 null,
                 Helper.CreateSprite("GatherMobileMedium.png"),
                 AbilityType.Special,
@@ -133,23 +170,171 @@ namespace DarkCodex
                 Helper.CreatePrerequisiteClassLevel(kineticist_class, 7, true),
                 Helper.CreateAddFacts(mobile_gathering_short_ab, mobile_gathering_long_ab));
             mobile_gathering_feat.Ranks = 1;
-            mobile_gathering_feat.IsClassFeature = true;
             Helper.AddFeats(mobile_gathering_feat);
 
             // make original gather ability visible for manual gathering and allow to extend buff3
             gather_original_ab.Hidden = false;
-            var hasBuff2 = gather_original_ab.GetComponent<AbilityEffectRunAction>().Actions.Actions[0] as Conditional;
-            Helper.AppendAndReplace(ref hasBuff2.ConditionsChecker.Conditions, Helper.CreateContextConditionHasBuff(buff3));
+            Helper.AppendAndReplace(ref gather_original_ab.GetComponent<AbilityEffectRunAction>().Actions.Actions, three2three);
+
         }
+
+        #region Helper
+
+        public static AbilityEffectRunAction step1_run_damage(PhysicalDamageForm p = 0, DamageEnergyType e = (DamageEnergyType)255, SavingThrowType save = SavingThrowType.Unknown, bool isAOE = false, bool half = false)
+        {
+            ContextDiceValue dice = Helper.CreateContextDiceValue(DiceType.D6, AbilityRankType.DamageDice, AbilityRankType.DamageBonus);
+
+            List<ContextAction> list = new List<ContextAction>(2);
+
+            if (p != 0)
+                list.Add(Helper.CreateContextActionDealDamage(p, dice, isAOE, isAOE, false, half));
+            if (e != (DamageEnergyType)255)
+                list.Add(Helper.CreateContextActionDealDamage(e, dice, isAOE, isAOE, false, half));
+
+            var runaction = Helper.CreateAbilityEffectRunAction(save, list.ToArray());
+
+            return runaction;
+        }
+
+        public static ContextRankConfig step2_rank_dice()
+        {
+            var rankdice = Helper.CreateContextRankConfig(
+                type: AbilityRankType.DamageDice,
+                baseValueType: ContextRankBaseValueType.FeatureRank,
+                feature: "93efbde2764b5504e98e6824cab3d27c".ToRef<BlueprintFeatureReference>()); //KineticBlastFeature
+            return rankdice;
+        }
+
+        public static ContextRankConfig step3_rank_bonus(bool half_bonus)
+        {
+            var rankdice = Helper.CreateContextRankConfig(
+                progression: half_bonus ? ContextRankProgression.Div2 : ContextRankProgression.AsIs,
+                type: AbilityRankType.DamageBonus,
+                baseValueType: ContextRankBaseValueType.CustomProperty,
+                stat: StatType.Constitution,
+                customProperty: "f897845bbbc008d4f9c1c4a03e22357a".ToRef<BlueprintUnitPropertyReference>()); //KineticistMainStatProperty
+            return rankdice;
+        }
+
+        public static ContextCalculateSharedValue step4_shared_value() // I think that's not used at all
+        {
+            return Helper.CreateContextCalculateSharedValue();
+        }
+
+        public static ContextCalculateAbilityParamsBasedOnClass step5dc()
+        {
+            var dc = new ContextCalculateAbilityParamsBasedOnClass();
+            dc.StatType = StatType.Dexterity;
+            dc.m_CharacterClass = "42a455d9ec1ad924d889272429eb8391".ToRef<BlueprintCharacterClassReference>(); //KineticistClass
+            return dc;
+        }
+
+        public static SpellDescriptorComponent step6_spell_description(SpellDescriptor descriptor)
+        {
+            return new SpellDescriptorComponent
+            {
+                Descriptor = descriptor
+            };
+        }
+
+        public static AbilityKineticist step7_burn(AbilityEffectRunAction run, int infusion = 0, int blast = 0, int talent = 0)
+        {
+            var list = new List<AbilityKineticist.DamageInfo>();
+            for (int i = 0; i < run.Actions.Actions.Length; i++)
+            {
+                var action = run.Actions.Actions[i] as ContextActionDealDamage;
+                if (action == null) continue;
+
+                list.Add(new AbilityKineticist.DamageInfo() { Value = action.Value, Type = action.DamageType, Half = action.Half });
+            }
+
+            var comp = new AbilityKineticist();
+            comp.InfusionBurnCost = infusion;
+            comp.BlastBurnCost = blast;
+            comp.WildTalentBurnCost = talent;
+            comp.CachedDamageInfo = list;
+            return comp;
+        }
+
+        public static AbilityShowIfCasterHasFact step8_feat(BlueprintUnitFact fact)
+        {
+            return Helper.CreateAbilityShowIfCasterHasFact(fact.ToRef());
+        }
+
+        public static AbilityDeliverProjectile step9_projectile(string projectile_guid, bool isPhysical, AbilityProjectileType type, float length, float width)
+        {
+            string weapon = isPhysical ? "65951e1195848844b8ab8f46d942f6e8" : "4d3265a5b9302ee4cab9c07adddb253f"; //KineticBlastPhysicalWeapon //KineticBlastEnergyWeapon
+
+            var projectile = Helper.CreateAbilityDeliverProjectile(
+                projectile_guid.ToRef<BlueprintProjectileReference>(),
+                type,
+                weapon.ToRef<BlueprintItemWeaponReference>(),
+                length.Feet(),
+                width.Feet());
+            return projectile;
+        }
+
+        public static AbilitySpawnFx step_sfx(AbilitySpawnFxTime time, string sfx_guid) //OnPrecastStart //OnStart
+        {
+            var sfx = new AbilitySpawnFx();
+            sfx.Time = time;
+            sfx.PrefabLink = new PrefabLink() { AssetId = sfx_guid };
+            return sfx;
+        }
+
+        //AbilityEffectRunAction - deals the actual damage; defines energy type (physical often half)
+        //ContextRankConfig - defines damage dice
+        //ContextRankConfig - defines bonus damage (half for energy)
+        //ContextCalculateSharedValue - ? (same for all?)
+        //ContextCalculateAbilityParamsBasedOnClass - defines primary stat (same for all)
+        //SpellDescriptorComponent - marks spell type
+        //AbilityKineticist - defines burn
+        //AbilityKineticist.CachedDamageInfo - shows damage preview (incl. energy type)
+        //AbilityShowIfCasterHasFact - feat requirement
+        //AbilityDeliverProjectile - which projectile to use
+        //AbilitySpawnFx - sfx x2
+
+        // 1) make BlueprintAbility
+        // 2) set m_Parent to XBlastBase
+        // 3) set SpellResistance
+        // 4) is step1_run_damage halved?
+        // 5) is step3_rank_bonus constitution bonus halved?
+
+        #endregion
     }
+
+    public static class Extension_Kin
+    {
+        //public static DamageEnergyType ToEnergy(this EnergyType type)
+        //{
+
+        //}
+    }
+
+    [Flags]
+    public enum EnergyType
+    {
+        None = 0,
+        Bludgeoning = 1,
+        Piercing = 2,
+        Slashing = 4,
+        Fire = 8,
+        Cold = 16,
+        Electric = 32,
+
+        Physical = 0x20000000,
+        Energy = 0x40000000,
+    }
+
 
     #region Patches
 
     /// <summary>
     /// Normal: Healing burn turns its non-lethal damage into actual damage.
     /// Patched: Healing burn heals non-lethal damage completely.
+    /// Correction: This might work differently than I thought. The HP bar is not consistant with the actual max HP. Needs investigation.
     /// </summary>
-    [HarmonyPatch(typeof(UnitPartKineticist), nameof(UnitPartKineticist.HealBurn))]
+    //[HarmonyPatch(typeof(UnitPartKineticist), nameof(UnitPartKineticist.HealBurn))]
     public class Patch_HealBurnDamage
     {
         public static void Postfix(int value, UnitPartKineticist __instance)
@@ -185,7 +370,7 @@ namespace DarkCodex
                 buffRank = 3;
 
             int value = KineticistUtils.CalculateGatherPowerBonus(kineticist.GatherPowerBaseValue, buffRank); // add increase from Supercharge
-            
+
             cost.IncreaseGatherPower(value); // apply value
 
             return false;
@@ -193,4 +378,19 @@ namespace DarkCodex
     }
 
     #endregion
+
+    //[HarmonyPatch(typeof(UnitCommands), nameof(UnitCommands.Run), typeof(UnitCommand))]
+    public class Patch_Debug1
+    {
+        public static void Prefix(UnitCommands __instance)
+        {
+            var cd = __instance.m_Owner?.CombatState?.Cooldown;
+            Helper.Print($"pre move={cd?.MoveAction} std={cd?.StandardAction}");
+        }
+        public static void Postfix(UnitCommands __instance)
+        {
+            var cd = __instance.m_Owner?.CombatState?.Cooldown;
+            Helper.Print($"post move={cd?.MoveAction} std={cd?.StandardAction}");
+        }
+    }
 }