+++ DarkCodex/Main.cs
@@ -11,14 +11,15 @@ using UnityModManagerNet;
 using HarmonyLib;
 using Guid = DarkCodex.GuidManager;
 using DarkCodex.Components;
+using Kingmaker.Modding;
+using Kingmaker.Achievements;
+using Kingmaker;
 
 namespace DarkCodex
 {
     public class Main
     {
         public static Harmony harmony;
-
-        public static bool COTWpresent = false;
 
         /// <summary>True if mod is enabled. Doesn't do anything right now.</summary>
         public static bool Enabled { get; set; } = true;
@@ -52,7 +53,7 @@ namespace DarkCodex
             GUILayout.Label("Legend: [F] This adds a feat. You still need to pick feats/talents for these effects. If you already picked these features, then they stay in effect regardless of the option above."
                 + "\n[*] Option is enabled/disabled immediately, without restart.");
 
-            //Checkbox(ref Settings.StateManager.State.slumberHDrestriction, "CotW - remove slumber HD restriction");
+            Checkbox(ref Settings.StateManager.State.allowAchievements, "[*] Allow achievements - enables achievements while mods are active and also set corresponding flag to future save files");
 
             GUILayout.Label("");
 
@@ -213,16 +214,22 @@ namespace DarkCodex
                 try
                 {
                     Helper.Print("Loading Dark Codex");
-                    //LoadSafe(CotW.modSlumber, Settings.StateManager.State.slumberHDrestriction);
+
                     LoadSafe(Kineticist.createKineticistBackground);
                     LoadSafe(Kineticist.createMobileGatheringFeat);
+                    LoadSafe(Kineticist.createExtraWildTalentFeat, true); //Settings.StateManager.State.
                     LoadSafe(Hexcrafter.fixProgression);
 
                     LoadSafe(Items.patchArrows);
+                    LoadSafe(Items.patchTerendelevScale);
+
                     LoadSafe(RestoreEndOfCombat.Activate);
 
-                    Helper.Print("Finished loading Dark Codex");
+                    harmony.Patch(
+                        original: AccessTools.PropertyGetter(typeof(OwlcatModificationsManager), "IsAnyModActive"),
+                        prefix: new HarmonyMethod(typeof(Patch_AllowAchievements), nameof(Patch_AllowAchievements.Prefix)));
 
+                    Helper.Print("Finished loading Dark Codex");
 #if DEBUG
                     Helper.PrintDebug("Running in debug.");
                     Guid.i.WriteAll();
@@ -319,36 +326,78 @@ namespace DarkCodex
             return false;
         }
 
-        #endregion
-
-        #region Special Patches
-
-        // Note: CallOfTheWild also patches this, but in a different manner. Not sure if there is potential in error. So far I haven't found any bugs in playtesting.
-        // Also, the enum will change, if the order of calls are changed, e.g. new mods are added. I think that's not a problem. I can't think of a reason why it would be saved in the save file. Needs further testing.
-        //[HarmonyLib.HarmonyPatch(typeof(EnumUtils), nameof(EnumUtils.GetMaxValue))] since this is a generic method, we need to patch this manually, see Main.Load
-        public static class Patch_ActivatableAbilityGroup
-        {
-            public static int ExtraGroups = 0;
-            public static bool GameAlreadyRunning = false;
-
-            ///<summary>Calls this to register a new group. Returns your new enum.</summary>
-            public static ActivatableAbilityGroup GetNewGroup()
-            {
-                if (GameAlreadyRunning)
-                    return 0;
-
-                ExtraGroups++;
-                Helper.PrintDebug("GetNewGroup new: " + (Enum.GetValues(typeof(ActivatableAbilityGroup)).Cast<int>().Max() + ExtraGroups).ToString());
-                return (ActivatableAbilityGroup)(Enum.GetValues(typeof(ActivatableAbilityGroup)).Cast<int>().Max() + ExtraGroups);
-            }
-
-            public static void Postfix(ref int __result)
-            {
-                __result += ExtraGroups;
-            }
-        }
+        // TODO: make patch helper?
 
         #endregion
 
     }
+
+    #region Special Patches
+
+    /// <summary>
+    /// [HarmonyPatch(typeof(OwlcatModificationsManager), nameof(OwlcatModificationsManager.IsAnyModActive), MethodType.Getter)]
+    /// Is manually patched. It crashes otherwise. Clears the 'has used mods before' flag and also pretends that no mods are active.
+    /// </summary>
+    public static class Patch_AllowAchievements
+    {
+        public static bool Prefix(ref bool __result)
+        {
+            if (!Settings.StateManager.State.allowAchievements)
+                return true;
+
+            if (Game.Instance?.Player != null)
+                Game.Instance.Player.ModsUser = false;
+            __result = false;
+            return false;
+        }
+    }
+
+    [HarmonyPatch(typeof(AchievementEntity), nameof(AchievementEntity.IsDisabled), MethodType.Getter)]
+    public static class Patch_AllowAchievements2
+    {
+        public static bool Prepare()
+        {
+#if DEBUG
+            return true;
+#else
+            return false;
+#endif
+        }
+
+        public static void Postfix(bool __result)
+        {
+            bool modactive = OwlcatModificationsManager.Instance.IsAnyModActive;
+            bool usermarked = Game.Instance.Player.ModsUser;
+
+            Helper.PrintDebug($"Achievements: dis={__result} modactive={modactive} usermarked={usermarked}");
+        }
+    }
+
+    // Note: CallOfTheWild also patches this, but in a different manner. Not sure if there is potential in error. So far I haven't found any bugs in playtesting.
+    // Also, the enum will change, if the order of calls are changed, e.g. new mods are added. I think that's not a problem. I can't think of a reason why it would be saved in the save file. Needs further testing.
+    //[HarmonyLib.HarmonyPatch(typeof(EnumUtils), nameof(EnumUtils.GetMaxValue))] since this is a generic method, we need to patch this manually, see Main.Load
+    public static class Patch_ActivatableAbilityGroup
+    {
+        public static int ExtraGroups = 0;
+        public static bool GameAlreadyRunning = false;
+
+        ///<summary>Calls this to register a new group. Returns your new enum.</summary>
+        public static ActivatableAbilityGroup GetNewGroup()
+        {
+            if (GameAlreadyRunning)
+                return 0;
+
+            ExtraGroups++;
+            Helper.PrintDebug("GetNewGroup new: " + (Enum.GetValues(typeof(ActivatableAbilityGroup)).Cast<int>().Max() + ExtraGroups).ToString());
+            return (ActivatableAbilityGroup)(Enum.GetValues(typeof(ActivatableAbilityGroup)).Cast<int>().Max() + ExtraGroups);
+        }
+
+        public static void Postfix(ref int __result)
+        {
+            __result += ExtraGroups;
+        }
+    }
+
+#endregion
+
 }